refactor(import): global로 모듈 이동
