--- src/page/lostfound/constants/lostfound.constant.ts
@@ -1,12 +1,0 @@
-const lostfound = {
-  lost: {
-    name: "분실",
-    type: "lost",
-  },
-  found: {
-    name: "습득",
-    type: "found",
-  },
-};
-
-export default lostfound;
