+++ src/page/timetable/layouts/TimeTableBox.tsx
@@ -3,7 +3,6 @@ import styled from "styled-components";
 import useDate from "@/hooks/useDate";
 import { useQueryClient } from "react-query";
 import KEY from "@/global/constants/key.constant";
-import IClassLevel from "@/global/types/classLevel.type";
 import TimeTableBar from "./TimeTableBar";
 import { useTimetableListQuery } from "../services/queries.service";
 import TimeTableCategory from "./TimeTableCategory";
@@ -11,16 +10,14 @@ import emptyTimetable from "../data/emptyTimetable";
 import emptyClassLevel from "../data/emptyClassLevel";
 
 const TimeTableBox = () => {
+  const queryClient = useQueryClient();
   const { weekdaysKOR: weekdays, getNowWeekDay, translateDay } = useDate();
   const [selectedDay, setSelectedDay] = React.useState<string>(
     getNowWeekDay({ type: "KOR" }),
   );
 
-  const [classLevel, setClassLevel] =
-    React.useState<IClassLevel>(emptyClassLevel);
+  const [classLevel, setClassLevel] = React.useState(emptyClassLevel);
   const [dayTimeTable, setDayTimeTable] = React.useState(emptyTimetable);
-
-  const queryClient = useQueryClient();
   const { data } = useTimetableListQuery(classLevel);
 
   React.useEffect(() => {
