+++ src/store/categories.store.ts
@@ -1,7 +1,8 @@
-import forum from "@/page/forum/constants/forum.constant";
+import forum from "@/global/constants/forum.constant";
+import PostCategoryType from "@/global/types/postCategory.type";
 import { atom } from "recoil";
 
-export const categoriesStore = atom<string>({
+export const categoriesStore = atom<PostCategoryType>({
   key: "categoriesStore",
   default: forum.category.all.type,
 });
