+++ src/store/forumType.store.ts
@@ -1,4 +1,4 @@
-import forum from "@/page/forum/constants/forum.constant";
+import forum from "@/global/constants/forum.constant";
 import { atom } from "recoil";
 
 export const forumFilterStore = atom<string>({
